+++ chef-repo/cookbooks/sakuravps/attributes/default.rb
@@ -1,2 +1,0 @@
-default['clamav']['clamd']['enabled'] = true
-default['clamav']['freshclam']['enabled'] = true