+++ chef-repo/site-cookbooks/eccube3/metadata.rb
@@ -1,9 +1,7 @@
-name             'lamp'
+name             'eccube3'
 maintainer       'LogicHeart'
 maintainer_email 'hiroyasu.yamada@logicheart.biz'
 license          'All rights reserved'
 description      'Installs/Configures CentOS6,Apache,PHP environment'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '1.0.0'
-depends          'yum-remi'
-depends          'mysql', '~> 5.6'