+++ chef-repo/site-cookbooks/eccube3/recipes/default.rb
@@ -1,0 +1,8 @@
+#
+# Cookbook Name:: eccube3
+# Recipe:: default
+#
+# Copyright 2015, hiroyasu55(LogicHeart)
+#
+# All rights reserved
+#